handlers/reader_handler.py
from appium.webdriver.common.appiumby import AppiumBy
from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.support import expected_conditions as EC


class ReaderHandler:
    def __init__(self, driver):
        self.driver = driver

    def turn_page(self, direction="forward"):
        """Turn the page forward or backward"""
        # TODO: Implement page turning functionality
        pass

    def get_current_page(self):
        """Get the current page number"""
        # TODO: Implement current page retrieval
        pass

    def get_reading_progress(self):
        """Get reading progress as percentage"""
        # TODO: Implement reading progress retrieval
        pass
