requirements.txt
selenium>=4.1.0
Appium-Python-Client>=2.6.1
